Merging changes from Murphy's scripts
hru_projection can now be set using a raster or feature class along with
a WKID or name.

Added function to compare remap files and rasters for missing values and
covden_win needed to be scaled by 0.1.

Snap/reference point can now be a float instead of just an int