Remap conversion now writes comment lines at the bottom of the files
ArcGIS 10.2 has a few bugs with comments lines in the ASCII remap files.
Comments must be fewer than 80 characters, most of the time they can't
be mixed in with the values or at the beginning (but not always?), and
there cannot be empty lines or a newline character at the end of the
file.  The conversion script tries to account for all of this.